+++ engines/stark/gfx/driver.cpp
@@ -45,23 +45,23 @@ Driver *Driver::create() {
 	Graphics::RendererType desiredRendererType = Graphics::parseRendererTypeCode(rendererConfig);
 	Graphics::RendererType matchingRendererType = Graphics::getBestMatchingAvailableRendererType(desiredRendererType);
 
-#if defined(USE_OPENGL_SHADERS) || defined(USE_GLES2)
+#if defined(USE_OPENGL_GAME) || defined(USE_OPENGL_SHADERS) || defined(USE_GLES2)
 	bool softRenderer = matchingRendererType == Graphics::kRendererTypeTinyGL;
 	if (!softRenderer) {
 		initGraphics3d(kOriginalWidth, kOriginalHeight);
 	} else {
 #endif
 		initGraphics(kOriginalWidth, kOriginalHeight, nullptr);
-#if defined(USE_OPENGL_SHADERS) || defined(USE_GLES2)
+#if defined(USE_OPENGL_GAME) || defined(USE_OPENGL_SHADERS) || defined(USE_GLES2)
 	}
 #endif
-
-	Driver *driver = nullptr;
 
 	if (matchingRendererType != desiredRendererType && desiredRendererType != Graphics::kRendererTypeDefault) {
 		// Display a warning if unable to use the desired renderer
 		warning("Unable to match a '%s' renderer", rendererConfig.c_str());
 	}
+
+	Driver *driver = nullptr;
 
 #if defined(USE_GLES2) || defined(USE_OPENGL_SHADERS)
 	bool backendCapableOpenGLShaders = g_system->hasFeature(OSystem::kFeatureOpenGLForGame) && OpenGLContext.shadersSupported;
@@ -69,19 +69,18 @@ Driver *Driver::create() {
 		driver = new OpenGLSDriver();
 	}
 #endif
-	if (matchingRendererType == Graphics::kRendererTypeTinyGL) {
-		//driver = new TinyGLDriver();
+#if defined(USE_OPENGL_GAME) && !defined(USE_GLES2)
+	bool backendCapableOpenGL = g_system->hasFeature(OSystem::kFeatureOpenGLForGame);
+	if (backendCapableOpenGL && matchingRendererType == Graphics::kRendererTypeOpenGL) {
+		driver = new OpenGLDriver();
 	}
+#endif
 
-	if (!driver) {
-		if (desiredRendererType != Graphics::kRendererTypeDefault) {
-			warning("Desired a '%s' renderer is not supported", rendererConfig.c_str());
-			GUI::displayErrorDialog(Common::U32String::format("Desired a '%s' renderer is not supported", rendererConfig.c_str()));
-		} else {
-			warning("No renderers have been found for this game");
-			GUI::displayErrorDialog(Common::U32String::format(_("No renderers have been found for this game")));
-		}
-	}
+	if (driver)
+		return driver;
+
+	warning("No renderers have been found for this game");
+	GUI::displayErrorDialog(Common::U32String::format(_("No renderers have been found for this game")));
 	return driver;
 }