+++ engines/stark/gfx/opengl.cpp
@@ -43,6 +43,7 @@ namespace Stark {
 namespace Gfx {
 
 OpenGLDriver::OpenGLDriver() {
+	_computeLights = true;
 }
 
 OpenGLDriver::~OpenGLDriver() {
@@ -60,6 +61,7 @@ void OpenGLDriver::init() {
 	glLoadIdentity();
 	glMatrixMode(GL_MODELVIEW);
 	glLoadIdentity();
+	glDisable(GL_LIGHTING);
 }
 
 void OpenGLDriver::setScreenViewport(bool noScaling) {
@@ -108,7 +110,6 @@ void OpenGLDriver::setupLights(const LightEntryArray &lights) {
 	assert(ambient->type == LightEntry::kAmbient); // The first light must be the ambient light
 
 	Math::Matrix4 viewMatrix = StarkScene->getViewMatrix();
-	Math::Matrix3 viewMatrixRot = viewMatrix.getRotation();
 
 	glMatrixMode(GL_MODELVIEW);
 	glLoadIdentity();
@@ -121,6 +122,7 @@ void OpenGLDriver::setupLights(const LightEntryArray &lights) {
 		GLfloat lightDir[] = { 0.0f, 0.0f, -1.0f };
 		GLfloat cutoff = 180.0f;
 		GLfloat spotExp = 0.0f;
+		GLfloat c_attenuation = 1.0f;
 		GLfloat l_attenuation = 0.0f;
 		GLfloat q_attenuation = 0.0f;
 
@@ -133,10 +135,9 @@ void OpenGLDriver::setupLights(const LightEntryArray &lights) {
 		Math::Vector4d eyePosition = viewMatrix * worldPosition;
 
 		Math::Vector3d worldDirection = l->direction;
-		Math::Vector3d eyeDirection = viewMatrixRot * worldDirection;
+		Math::Vector3d eyeDirection = viewMatrix.getRotation() * worldDirection;
 		eyeDirection.normalize();
 
-		glDisable(GL_LIGHT0 + i);
 		switch (l->type) {
 			case LightEntry::kPoint:
 				lightColor[0] = (GLfloat)l->color.x();
@@ -150,9 +151,9 @@ void OpenGLDriver::setupLights(const LightEntryArray &lights) {
 				lightColor[0] = (GLfloat)l->color.x();
 				lightColor[1] = (GLfloat)l->color.y();
 				lightColor[2] = (GLfloat)l->color.z();
-				lightPos[0] = -(GLfloat)eyeDirection.x();
-				lightPos[1] = -(GLfloat)eyeDirection.y();
-				lightPos[2] = -(GLfloat)eyeDirection.z();
+				lightPos[0] = (GLfloat)-eyeDirection.x();
+				lightPos[1] = (GLfloat)-eyeDirection.y();
+				lightPos[2] = (GLfloat)-eyeDirection.z();
 				lightPos[3] = 0;
 				break;
 			case LightEntry::kSpot:
@@ -162,19 +163,15 @@ void OpenGLDriver::setupLights(const LightEntryArray &lights) {
 				lightPos[0] = (GLfloat)eyePosition.x();
 				lightPos[1] = (GLfloat)eyePosition.y();
 				lightPos[2] = (GLfloat)eyePosition.z();
-				lightDir[0] = -(GLfloat)eyeDirection.x();
-				lightDir[1] = -(GLfloat)eyeDirection.y();
-				lightDir[2] = -(GLfloat)eyeDirection.z();
-				// FIXME
-				l_attenuation = 0.0000001f;
-				q_attenuation = 0.0000001f;
-				//spotExp = 0.0f;
-				//cutoff = (l->outerConeAngle.getDegrees() + l->innerConeAngle.getDegrees()) / 2.0f;
+				lightDir[0] = (GLfloat)eyeDirection.x();
+				lightDir[1] = (GLfloat)eyeDirection.y();
+				lightDir[2] = (GLfloat)eyeDirection.z();
+				cutoff = (l->outerConeAngle.getDegrees() + l->innerConeAngle.getDegrees()) / 2.26f;
 				break;
 			case LightEntry::kAmbient:
-				ambientColor[0] = (GLfloat)l->color.x();
-				ambientColor[1] = (GLfloat)l->color.y();
-				ambientColor[2] = (GLfloat)l->color.z();
+				lightColor[0] = (GLfloat)l->color.x();
+				lightColor[1] = (GLfloat)l->color.y();
+				lightColor[2] = (GLfloat)l->color.z();
 				break;
 			default:
 				break;
@@ -186,6 +183,7 @@ void OpenGLDriver::setupLights(const LightEntryArray &lights) {
 		glLightfv(GL_LIGHT0 + i, GL_SPOT_DIRECTION, lightDir);
 		glLightf(GL_LIGHT0 + i, GL_SPOT_EXPONENT, spotExp);
 		glLightf(GL_LIGHT0 + i, GL_SPOT_CUTOFF, cutoff);
+		glLightf(GL_LIGHT0 + i, GL_CONSTANT_ATTENUATION, c_attenuation);
 		glLightf(GL_LIGHT0 + i, GL_LINEAR_ATTENUATION, l_attenuation);
 		glLightf(GL_LIGHT0 + i, GL_QUADRATIC_ATTENUATION, q_attenuation);
 		glEnable(GL_LIGHT0 + i);
@@ -207,6 +205,10 @@ Texture *OpenGLDriver::createTexture(const Graphics::Surface *surface, const byt
 	return texture;
 }
 
+Texture *OpenGLDriver::createBitmap(const Graphics::Surface *surface, const byte *palette) {
+	return createTexture(surface, palette);
+}
+
 VisualActor *OpenGLDriver::createActorRenderer() {
 	return new OpenGLActorRenderer(this);
 }
@@ -224,6 +226,11 @@ FadeRenderer *OpenGLDriver::createFadeRenderer() {
 }
 
 void OpenGLDriver::start2DMode() {
+#if defined(USE_OPENGL_SHADERS)
+	// The ShaderSurfaceRenderer sets an array buffer which conflict with fixed pipeline rendering
+	glBindBuffer(GL_ARRAY_BUFFER, 0);
+#endif // defined(USE_OPENGL_SHADERS)
+
 	// Enable alpha blending
 	glEnable(GL_BLEND);
 	//glBlendEquation(GL_FUNC_ADD); // It's the default
@@ -235,7 +242,8 @@ void OpenGLDriver::start2DMode() {
 
 	glDisable(GL_DEPTH_TEST);
 	glDepthMask(GL_FALSE);
-	glDisable(GL_LIGHTING);
+	if (!_computeLights)
+		glDisable(GL_LIGHTING);
 }
 
 void OpenGLDriver::end2DMode() {
@@ -244,10 +252,14 @@ void OpenGLDriver::end2DMode() {
 
 	glEnable(GL_DEPTH_TEST);
 	glDepthMask(GL_TRUE);
-	glEnable(GL_LIGHTING);
 }
 
 void OpenGLDriver::set3DMode() {
+#if defined(USE_OPENGL_SHADERS)
+	// The ShaderSurfaceRenderer sets an array buffer which conflict with fixed pipeline rendering
+	glBindBuffer(GL_ARRAY_BUFFER, 0);
+#endif // defined(USE_OPENGL_SHADERS)
+
 	glEnable(GL_DEPTH_TEST);
 	glDepthFunc(GL_LESS);
 
@@ -256,6 +268,13 @@ void OpenGLDriver::set3DMode() {
 	glBlendFunc(GL_SRC_ALPHA, GL_ONE_MINUS_SRC_ALPHA);
 	glStencilFunc(GL_EQUAL, 0, 0xFF);
 	glStencilOp(GL_KEEP, GL_KEEP, GL_INCR);
+
+	if (!_computeLights)
+		glEnable(GL_LIGHTING);
+}
+
+bool OpenGLDriver::computeLightsEnabled() {
+	return _computeLights;
 }
 
 Common::Rect OpenGLDriver::getViewport() const {