+++ engines/stark/gfx/openglactor.cpp
@@ -66,17 +66,31 @@ void OpenGLActorRenderer::render(const Math::Vector3d &position, float direction
 	Math::Vector3d lightDirection;
 
 	_gfx->set3DMode();
-	_gfx->setupLights(lights);
+	if (!_gfx->computeLightsEnabled())
+		_gfx->setupLights(lights);
 
 	Math::Matrix4 model = getModelMatrix(position, direction);
 	Math::Matrix4 view = StarkScene->getViewMatrix();
 	Math::Matrix4 projection = StarkScene->getProjectionMatrix();
 
 	Math::Matrix4 modelViewMatrix = view * model;
-	modelViewMatrix.transpose(); // OpenGL expects matrices transposed when compared to ScummVM's
+	modelViewMatrix.transpose(); // OpenGL expects matrices transposed
+	glMatrixMode(GL_MODELVIEW);
+	glLoadMatrixf(modelViewMatrix.getData());
 
 	Math::Matrix4 projectionMatrix = projection;
-	projectionMatrix.transpose(); // OpenGL expects matrices transposed when compared to ScummVM's
+	projectionMatrix.transpose(); // OpenGL expects matrices transposed
+	glMatrixMode(GL_PROJECTION);
+	glLoadMatrixf(projectionMatrix.getData());
+
+	Math::Matrix4 normalMatrix;
+	if (_gfx->computeLightsEnabled()) {
+		projectionMatrix.transpose();
+		modelViewMatrix.transpose();
+
+		normalMatrix = modelViewMatrix;
+		normalMatrix.invertAffineOrthonormal();
+	}
 
 	Math::Matrix4 mvp;
 	if (drawShadow) {
@@ -87,40 +101,43 @@ void OpenGLActorRenderer::render(const Math::Vector3d &position, float direction
 		lightDirection = getShadowLightDirection(lights, position, modelInverse.getRotation());
 	}
 
-	glMatrixMode(GL_PROJECTION);
-	glLoadMatrixf(projectionMatrix.getData());
-
-	glMatrixMode(GL_MODELVIEW);
-	glLoadMatrixf(modelViewMatrix.getData());
-
 	glEnable(GL_TEXTURE_2D);
 
 	Common::Array<Face *> faces = _model->getFaces();
 	Common::Array<Material *> mats = _model->getMaterials();
 	const Common::Array<BoneNode *> &bones = _model->getBones();
 
-	glColorMaterial(GL_FRONT_AND_BACK, GL_AMBIENT_AND_DIFFUSE);
-	glEnable(GL_COLOR_MATERIAL);
+	if (!_gfx->computeLightsEnabled()) {
+		glColorMaterial(GL_FRONT_AND_BACK, GL_DIFFUSE);
+		glEnable(GL_COLOR_MATERIAL);
+	}
+
 	for (Common::Array<Face *>::const_iterator face = faces.begin(); face != faces.end(); ++face) {
 		const Material *material = mats[(*face)->materialId];
+		Math::Vector3d color;
 		const Gfx::Texture *tex = resolveTexture(material);
-		if (tex) {
-			tex->bind();
-			glColor3f(1.0f, 1.0f, 1.0f);
-		} else {
-			glBindTexture(GL_TEXTURE_2D, 0);
-			glColor3f(material->r, material->g, material->b);
-		}
-
 		auto vertexIndices = _faceEBO[*face];
 		auto numVertexIndices = (*face)->vertexIndices.size();
 		for (uint32 i = 0; i < numVertexIndices; i++) {
+			if (tex) {
+				tex->bind();
+				if (_gfx->computeLightsEnabled())
+					color = Math::Vector3d(1.0f, 1.0f, 1.0f);
+				else
+					glColor3f(1.0f, 1.0f, 1.0f);
+			} else {
+				glBindTexture(GL_TEXTURE_2D, 0);
+				if (_gfx->computeLightsEnabled())
+					color = Math::Vector3d(material->r, material->g, material->b);
+				else
+					glColor3f(material->r, material->g, material->b);
+			}
 			uint32 index = vertexIndices[i];
 			auto vertex = _faceVBO[index];
 			uint32 bone1 = vertex.bone1;
 			uint32 bone2 = vertex.bone2;
-			Math::Vector3d position1 = vertex.pos1;
-			Math::Vector3d position2 = vertex.pos2;
+			Math::Vector3d position1 = Math::Vector3d(vertex.pos1x, vertex.pos1y, vertex.pos1z);
+			Math::Vector3d position2 = Math::Vector3d(vertex.pos2x, vertex.pos2y, vertex.pos2z);
 			Math::Vector3d bone1Position = Math::Vector3d(bones[bone1]->_animPos.x(),
 														  bones[bone1]->_animPos.y(),
 														  bones[bone1]->_animPos.z());
@@ -136,27 +153,38 @@ void OpenGLActorRenderer::render(const Math::Vector3d &position, float direction
 														  bones[bone2]->_animRot.z(),
 														  bones[bone2]->_animRot.w());
 			float boneWeight = vertex.boneWeight;
-			Math::Vector3d normal = vertex.normal;
+			Math::Vector3d normal = Math::Vector3d(vertex.normalx, vertex.normaly, vertex.normalz);
 
 			// Compute the vertex position in eye-space
 			bone1Rotation.transform(position1);
 			position1 += bone1Position;
 			bone2Rotation.transform(position2);
 			position2 += bone2Position;
-			Math::Vector3d modelPosition = position2 * (1.0 - boneWeight) + position1 * boneWeight;
+			Math::Vector3d modelPosition = Math::Vector3d::interpolate(position2, position1, boneWeight);
 			vertex.x = modelPosition.x();
 			vertex.y = modelPosition.y();
 			vertex.z = modelPosition.z();
-
+			Math::Vector4d modelEyePosition;
+			if (_gfx->computeLightsEnabled()) {
+				modelEyePosition = modelViewMatrix * Math::Vector4d(modelPosition.x(),
+																		       modelPosition.y(),
+																		       modelPosition.z(),
+																		       1.0);
+			}
 			// Compute the vertex normal in eye-space
 			Math::Vector3d n1 = normal;
 			bone1Rotation.transform(n1);
 			Math::Vector3d n2 = normal;
 			bone2Rotation.transform(n2);
-			Math::Vector3d modelNormal = Math::Vector3d(n2 * (1.0 - boneWeight) + n1 * boneWeight).getNormalized();
+			Math::Vector3d modelNormal = Math::Vector3d(Math::Vector3d::interpolate(n2, n1, boneWeight)).getNormalized();
 			vertex.nx = modelNormal.x();
 			vertex.ny = modelNormal.y();
 			vertex.nz = modelNormal.z();
+			Math::Vector3d modelEyeNormal;
+			if (_gfx->computeLightsEnabled()) {
+				modelEyeNormal = normalMatrix.getRotation() * modelNormal;
+				modelEyeNormal.normalize();
+			}
 
 			if (drawShadow) {
 				Math::Vector3d shadowPosition = modelPosition + lightDirection * (-modelPosition.y() / lightDirection.y());
@@ -165,33 +193,100 @@ void OpenGLActorRenderer::render(const Math::Vector3d &position, float direction
 				vertex.sz = shadowPosition.z();
 			}
 
+			if (_gfx->computeLightsEnabled()) {
+				static const uint maxLights = 10;
+
+				assert(lights.size() >= 1);
+				assert(lights.size() <= maxLights);
+
+				const LightEntry *ambient = lights[0];
+				assert(ambient->type == LightEntry::kAmbient); // The first light must be the ambient light
+
+				Math::Vector3d lightColor = ambient->color;
+
+				for (uint li = 0; li < lights.size() - 1; li++) {
+					const LightEntry *l = lights[li + 1];
+
+					switch (l->type) {
+						case LightEntry::kPoint: {
+							Math::Vector3d vertexToLight = l->eyePosition.getXYZ() - modelEyePosition.getXYZ();
+
+							float dist = vertexToLight.length();
+							vertexToLight.normalize();
+							float attn = CLIP((l->falloffFar - dist) / MAX(0.001f,  l->falloffFar - l->falloffNear), 0.0f, 1.0f);
+							float incidence = MAX(0.0f, Math::Vector3d::dotProduct(modelEyeNormal, vertexToLight));
+							lightColor += l->color * attn * incidence;
+							break;
+						}
+						case LightEntry::kDirectional: {
+							float incidence = MAX(0.0f, Math::Vector3d::dotProduct(modelEyeNormal, -l->eyeDirection));
+							lightColor += (l->color * incidence);
+							break;
+						}
+						case LightEntry::kSpot: {
+							Math::Vector3d vertexToLight = l->eyePosition.getXYZ() - modelEyePosition.getXYZ();
+
+							float dist = vertexToLight.length();
+							float attn = CLIP((l->falloffFar - dist) / MAX(0.001f, l->falloffFar - l->falloffNear), 0.0f, 1.0f);
+
+							vertexToLight.normalize();
+							float incidence = MAX(0.0f, modelEyeNormal.dotProduct(vertexToLight));
+
+							float cosAngle = MAX(0.0f, vertexToLight.dotProduct(-l->eyeDirection));
+							float cone = CLIP((cosAngle - l->innerConeAngle.getCosine()) / MAX(0.001f, l->outerConeAngle.getCosine() - l->innerConeAngle.getCosine()), 0.0f, 1.0f);
+
+							lightColor += l->color * attn * incidence * cone;
+							break;
+						}
+						default:
+							break;
+					}
+				}
+
+				lightColor.x() = CLIP(lightColor.x(), 0.0f, 1.0f);
+				lightColor.y() = CLIP(lightColor.y(), 0.0f, 1.0f);
+				lightColor.z() = CLIP(lightColor.z(), 0.0f, 1.0f);
+				color = color * lightColor;
+				vertex.r = color.x();
+				vertex.g = color.y();
+				vertex.b = color.z();
+			}
+
 			_faceVBO[index] = vertex;
 		}
 
 		glEnableClientState(GL_VERTEX_ARRAY);
-		glEnableClientState(GL_TEXTURE_COORD_ARRAY);
+		if (_gfx->computeLightsEnabled())
+			glEnableClientState(GL_COLOR_ARRAY);
+		if (tex)
+			glEnableClientState(GL_TEXTURE_COORD_ARRAY);
 		glEnableClientState(GL_NORMAL_ARRAY);
 
 		glVertexPointer(3, GL_FLOAT, sizeof(ActorVertex), &_faceVBO[0].x);
 		if (tex)
 			glTexCoordPointer(2, GL_FLOAT, sizeof(ActorVertex), &_faceVBO[0].texS);
 		glNormalPointer(GL_FLOAT, sizeof(ActorVertex), &_faceVBO[0].nx);
+		if (_gfx->computeLightsEnabled())
+			glColorPointer(3, GL_FLOAT, sizeof(ActorVertex), &_faceVBO[0].r);
 
 		glDrawElements(GL_TRIANGLES, numVertexIndices, GL_UNSIGNED_INT, vertexIndices);
 
 		glDisableClientState(GL_VERTEX_ARRAY);
-		glDisableClientState(GL_NORMAL_ARRAY);
+		if (_gfx->computeLightsEnabled())
+			glDisableClientState(GL_COLOR_ARRAY);
 		glDisableClientState(GL_TEXTURE_COORD_ARRAY);
-
+		glDisableClientState(GL_NORMAL_ARRAY);
 	}
-	glDisable(GL_COLOR_MATERIAL);
+	if (!_gfx->computeLightsEnabled())
+		glDisable(GL_COLOR_MATERIAL);
 
 
 	if (drawShadow) {
 		glEnable(GL_BLEND);
 		glEnable(GL_STENCIL_TEST);
 		glDisable(GL_TEXTURE_2D);
-		glDisable(GL_LIGHTING);
+		if (!_gfx->computeLightsEnabled())
+			glDisable(GL_LIGHTING);
 
 		glMatrixMode(GL_PROJECTION);
 		glLoadMatrixf(projectionMatrix.getData());
@@ -213,7 +308,8 @@ void OpenGLActorRenderer::render(const Math::Vector3d &position, float direction
 			glDisableClientState(GL_VERTEX_ARRAY);
 		}
 
-		glEnable(GL_LIGHTING);
+		if (!_gfx->computeLightsEnabled())
+			glEnable(GL_LIGHTING);
 		glEnable(GL_TEXTURE_2D);
 		glDisable(GL_BLEND);
 		glDisable(GL_STENCIL_TEST);
@@ -247,12 +343,18 @@ ActorVertex *OpenGLActorRenderer::createModelVBO(const Model *model) {
 	// Build a vertex array
 	int i = 0;
 	for (Common::Array<VertNode *>::const_iterator tri = modelVertices.begin(); tri != modelVertices.end(); ++tri, i++) {
-		vertices[i].pos1 = Math::Vector3d((*tri)->_pos1.x(), (*tri)->_pos1.y(), (*tri)->_pos1.z());
-		vertices[i].pos2 = Math::Vector3d((*tri)->_pos2.x(), (*tri)->_pos2.y(), (*tri)->_pos2.z());
+		vertices[i].pos1x = (*tri)->_pos1.x();
+		vertices[i].pos1y = (*tri)->_pos1.y();
+		vertices[i].pos1z = (*tri)->_pos1.z();
+		vertices[i].pos2x = (*tri)->_pos2.x();
+		vertices[i].pos2y = (*tri)->_pos2.y();
+		vertices[i].pos2z = (*tri)->_pos2.z();
 		vertices[i].bone1 = (*tri)->_bone1;
 		vertices[i].bone2 = (*tri)->_bone2;
 		vertices[i].boneWeight = (*tri)->_boneWeight;
-		vertices[i].normal = Math::Vector3d((*tri)->_normal.x(), (*tri)->_normal.y(), (*tri)->_normal.z());
+		vertices[i].normalx = (*tri)->_normal.x();
+		vertices[i].normaly = (*tri)->_normal.y();
+		vertices[i].normalz = (*tri)->_normal.z();
 		vertices[i].texS = -(*tri)->_texS;
 		vertices[i].texT = (*tri)->_texT;
 	}